backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gunicorn>=21.2.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
psycopg[binary]>=3.2.0
alembic>=1.13.1

# Redis
redis>=5.0.1
hiredis>=2.3.2

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0

# ML and Data Science
scikit-learn>=1.4.0
xgboost>=2.0.3
tensorflow>=2.16.0
shap>=0.44.1
numpy>=1.26.0
pandas>=2.1.0
scipy>=1.11.0
joblib>=1.3.0

# Medical Imaging
pydicom>=2.4.4
nibabel>=5.2.0
cryptography>=41.0.0

# Background Tasks
celery>=5.3.6
flower>=2.0.1

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
aiofiles>=23.2.1
